cms-admin: Move ActionLogCompare and ActionLogShowVersion to components folder
Both components are dumb presentational views (no GraphQL query, no
workflow state) and will be reused by the per-record, per-entity-type,
and cross-entity action log views. Co-locating them with the other
shared building blocks (UserCell, ActionChip, DiffHeader, DiffViewer,
ActionLogHeader) keeps the components folder as the single home for
reusable dumb action log pieces.

Updates the public exports in src/index.ts and all dialog importers.
