+++ packages/admin/cms-admin/src/actionLog/components/actionLogShowVersion/ActionLogShowVersion.styles.ts
