+++ packages/admin/cms-admin/src/index.ts
@@ -1,10 +1,10 @@
 export { ActionLogButton } from "./actionLog/actionLogButton/ActionLogButton";
-export { ActionLogCompare, actionLogCompareFragment } from "./actionLog/actionLogCompare/ActionLogCompare";
 export { ActionLogDialog } from "./actionLog/actionLogDialog/ActionLogDialog";
 export { ActionLogGrid, actionLogGridFragment } from "./actionLog/actionLogGrid/ActionLogGrid";
 export { ActionLogsGrid, actionLogsGridFragment } from "./actionLog/actionLogs/actionLogsGrid/ActionLogsGrid";
 export { ActionLogsPage } from "./actionLog/actionLogs/actionLogsPage/ActionLogsPage";
-export { ActionLogShowVersion, actionLogShowVersionFragment } from "./actionLog/actionLogShowVersion/ActionLogShowVersion";
+export { ActionLogCompare, actionLogCompareFragment } from "./actionLog/components/actionLogCompare/ActionLogCompare";
+export { ActionLogShowVersion, actionLogShowVersionFragment } from "./actionLog/components/actionLogShowVersion/ActionLogShowVersion";
 export { DiffHeader, type DiffHeaderProps } from "./actionLog/components/diffHeader/DiffHeader";
 export { DiffViewer, type DiffViewerProps } from "./actionLog/components/diffViewer/DiffViewer";
 export { ActionLogHeader, type ActionLogHeaderProps } from "./actionLog/components/header/ActionLogHeader";
